services/model-serving/requirements.txt
accelerate==1.0.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
datasets==3.0.1
dill==0.3.8
fastapi==0.115.0
fsspec==2024.6.1
FlagEmbedding==1.2.11
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.4
multiprocess==0.70.16
MarkupSafe==3.0.0
numpy==2.0.2
pandas==2.2.3
peft==0.13.1
pillow==10.4.0
pip==24.2
pyarrow==17.0.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
setuptools==75.1.0
scikit-learn==1.5.2
sentence-transformers==3.1.1
starlette==0.38.6
tokenizers==0.20.0
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
transformers==4.45.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.0
wheel==0.44.0
xxhash==3.5.0
